+++ CHANGELOG.md
@@ -9,7 +9,8 @@ Any contributor making changes to the project must record their changes in this 
 ## TOC
 
 - [Unreleased](#unreleased)
-- [Release 0.3.1 (Unreleased)](#031-unreleased)
+- [Release 0.3.2 (Unreleased)](#031-unreleased)
+- [Release 0.3.1 (2026-09-16)](#031-2026-09-16)
 - [Release 0.3.0 (2026-09-06)](#030-2026-09-06)
 - [Release 0.2.0 (2026-08-10)](#020-2026-08-10)
 
@@ -37,7 +38,27 @@ None
 
 ---
 
-### 0.3.1 (Unreleased)
+### 0.3.2 (Unreleased)
+
+#### Fixes:
+
+None
+
+#### Optimizations:
+
+None
+
+#### Features:
+
+None
+
+#### **BREAKING CHANGES** (API CHANGES):
+
+None
+
+---
+
+### 0.3.1 (2026-09-16)
 
 #### Fixes: